requirements.txt
pydantic
flask
memory-profiler
paho-mqtt
colorlog